add authentication and auth middleware

+++ backend/controllers/feed.js
@@ -3,11 +3,24 @@ const path = require("path")
 const { validationResult } = require("express-validator/check")
 
 const Post = require("../models/post")
+const User = require("../models/user")
 
 exports.getPosts = (req, res, next) => {
+	const currentPage = req.query.page || 1
+	const perPage = 2
+	let totalItems
 	Post.find()
+		.countDocuments()
+		.then((count) => {
+			totalItems = count
+			return Post.find()
+				.skip((currentPage - 1) * perPage)
+				.limit(perPage)
+		})
 		.then((posts) => {
-			res.status(200).json({ message: "Fetched posts successfully.", posts })
+			res
+				.status(200)
+				.json({ message: "Fetched posts successfully.", posts, totalItems })
 		})
 		.catch((err) => {
 			if (!err.statusCode) {
@@ -34,19 +47,30 @@ exports.createPost = (req, res, next) => {
 
 	const { title, content } = req.body
 	const imageUrl = req.file.path.replace("\\", "/")
-	// Create post in db
+	let creator
+
 	const post = new Post({
 		title,
 		content,
 		imageUrl,
-		creator: { name: "Daniel" },
+		creator: req.userId,
 	})
+
 	post
 		.save()
 		.then((result) => {
+			return User.findById(req.userId)
+		})
+		.then((user) => {
+			creator = user
+			user.posts.push(post)
+			return user.save()
+		})
+		.then((result) => {
 			res.status(201).json({
 				message: "Post created successfully",
-				post: result,
+				post: post,
+				creator: { _id: creator._id, name: creator.name },
 			})
 		})
 		.catch((err) => {
@@ -107,6 +131,11 @@ exports.updatePost = (req, res, next) => {
 				error.statusCode = 404
 				throw error
 			}
+			if (post.creator.toString() !== req.userId) {
+				const error = new Error("Not authorized!")
+				error.statusCode = 403
+				throw error
+			}
 			if (imageUrl !== post.imageUrl) {
 				clearImage(post.imageUrl)
 			}
@@ -143,8 +172,21 @@ exports.deletePost = (req, res, next) => {
 				throw error
 			}
 
+			if (post.creator.toString() !== req.userId) {
+				const error = new Error("Not authorized!")
+				error.statusCode = 403
+				throw error
+			}
+
 			clearImage(post.imageUrl)
 			return Post.findByIdAndRemove(postId)
+		})
+		.then((result) => {
+			return User.findById(req.userId)
+		})
+		.then((user) => {
+			user.posts.pull(postId)
+			return user.save()
 		})
 		.then((result) => {
 			console.log(result)
